Refactor movey-login logic, refactor movey tests
