+++ language/tools/move-cli/src/base/mod.rs
@@ -7,6 +7,7 @@ pub mod disassemble;
 pub mod docgen;
 pub mod errmap;
 pub mod info;
+pub mod movey_login;
 pub mod new;
 pub mod prove;
 pub mod test;
