+++ language/tools/move-cli/src/base/movey_login.rs
@@ -3,100 +3,85 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use anyhow::{bail, Result};
-use move_command_line_common::{env::MOVE_HOME, movey::MOVEY_URL};
+use clap::Parser;
+use move_command_line_common::{env::MOVE_HOME, movey_constants::MOVEY_URL};
 use std::{fs, fs::File, io, path::PathBuf};
 use toml_edit::easy::{map::Map, Value};
 
-pub const MOVEY_API_KEY_PATH: &str = "/movey_api_key.toml";
+pub const MOVEY_CREDENTIAL_PATH: &str = "/movey_credential.toml";
 
-pub struct TestMode {
-    pub test_path: String,
-}
+#[derive(Parser)]
+#[clap(name = "movey-login")]
+pub struct MoveyLogin;
 
-pub fn handle_movey_login_commands(test_path: Option<String>) -> Result<()> {
-    println!(
-        "Please paste the API Token found on {}/settings/tokens below",
-        MOVEY_URL
-    );
-    let mut line = String::new();
-    loop {
-        match io::stdin().read_line(&mut line) {
-            Ok(_) => {
-                if let Some('\n') = line.chars().next_back() {
-                    line.pop();
+impl MoveyLogin {
+    pub fn execute(self) -> Result<()> {
+        println!(
+            "Please paste the API Token found on {}/settings/tokens below",
+            MOVEY_URL
+        );
+        let mut line = String::new();
+        loop {
+            match io::stdin().read_line(&mut line) {
+                Ok(_) => {
+                    line = line.trim().to_string();
+                    if !line.is_empty() {
+                        break;
+                    }
+                    println!("Invalid API Token. Try again!");
                 }
-                if let Some('\r') = line.chars().next_back() {
-                    line.pop();
+                Err(err) => {
+                    bail!("Error reading file: {}", err);
                 }
-                if !line.is_empty() {
-                    break;
-                }
-                println!("Invalid API Token. Try again!");
-            }
-            Err(err) => {
-                bail!("Error reading file: {}", err);
             }
         }
+        Self::save_credential(line, MOVE_HOME.clone())?;
+        println!("Token for Movey saved.");
+        Ok(())
     }
-    let mut test_mode: Option<TestMode> = None;
-    if let Some(path) = test_path {
-        test_mode = Some(TestMode { test_path: path });
-    }
-    save_credential(line, test_mode)?;
-    println!("Token for Movey saved.");
-    Ok(())
-}
 
-pub fn save_credential(token: String, test_mode: Option<TestMode>) -> Result<()> {
-    let mut move_home;
-    if let Some(test_mode) = test_mode {
-        move_home = std::env::var("TEST_MOVE_HOME").unwrap();
-        if !test_mode.test_path.is_empty() {
-            move_home.push_str(&test_mode.test_path);
+    pub fn save_credential(token: String, move_home: String) -> Result<()> {
+        fs::create_dir_all(&move_home)?;
+        let credential_path = move_home + MOVEY_CREDENTIAL_PATH;
+        let credential_file = PathBuf::from(&credential_path);
+        if !credential_file.exists() {
+            let credential_file = File::create(&credential_path)?;
+            set_permissions(&credential_file, 0o600)?;
         }
-    } else {
-        move_home = MOVE_HOME.clone();
-    }
-    fs::create_dir_all(&move_home)?;
-    let credential_path = move_home + MOVEY_API_KEY_PATH;
-    let credential_file = PathBuf::from(&credential_path);
-    if !credential_file.exists() {
-        File::create(&credential_path)?;
-    }
 
-    let old_contents: String;
-    match fs::read_to_string(&credential_path) {
-        Ok(contents) => {
-            old_contents = contents;
+        let old_contents: String;
+        match fs::read_to_string(&credential_path) {
+            Ok(contents) => {
+                old_contents = contents;
+            }
+            Err(error) => bail!("Error reading input: {}", error),
         }
-        Err(error) => bail!("Error reading input: {}", error),
-    }
-    let mut toml: Value = old_contents
-        .parse()
-        .map_err(|e| anyhow::Error::from(e).context("could not parse input as TOML"))?;
+        let mut toml: Value = old_contents
+            .parse()
+            .map_err(|e| anyhow::Error::from(e).context("could not parse input as TOML"))?;
 
-    if let Some(registry) = toml.as_table_mut().unwrap().get_mut("registry") {
-        if let Some(toml_token) = registry.as_table_mut().unwrap().get_mut("token") {
-            *toml_token = Value::String(token);
+        // only update token key, keep the rest of the file intact
+        if let Some(registry) = toml.as_table_mut().unwrap().get_mut("registry") {
+            if let Some(toml_token) = registry.as_table_mut().unwrap().get_mut("token") {
+                *toml_token = Value::String(token);
+            } else {
+                registry
+                    .as_table_mut()
+                    .unwrap()
+                    .insert(String::from("token"), Value::String(token));
+            }
         } else {
-            registry
-                .as_table_mut()
+            let mut value = Map::new();
+            value.insert(String::from("token"), Value::String(token));
+            toml.as_table_mut()
                 .unwrap()
-                .insert(String::from("token"), Value::String(token));
+                .insert(String::from("registry"), Value::Table(value));
         }
-    } else {
-        let mut value = Map::new();
-        value.insert(String::from("token"), Value::String(token));
-        toml.as_table_mut()
-            .unwrap()
-            .insert(String::from("registry"), Value::Table(value));
-    }
 
-    let new_contents = toml.to_string();
-    fs::write(credential_file, new_contents).expect("Unable to write file");
-    let file = File::open(&credential_path)?;
-    set_permissions(&file, 0o600)?;
-    Ok(())
+        let new_contents = toml.to_string();
+        fs::write(credential_file, new_contents).expect("Unable to write file");
+        Ok(())
+    }
 }
 
 #[cfg(unix)]
@@ -123,13 +108,12 @@ mod tests {
     fn setup_move_home(test_path: &str) -> (String, String) {
         let cwd = env::current_dir().unwrap();
         let mut move_home: String = String::from(cwd.to_string_lossy());
-        env::set_var("TEST_MOVE_HOME", &move_home);
         if !test_path.is_empty() {
             move_home.push_str(&test_path);
         } else {
             move_home.push_str("/test");
         }
-        let credential_path = move_home.clone() + MOVEY_API_KEY_PATH;
+        let credential_path = move_home.clone() + MOVEY_CREDENTIAL_PATH;
         (move_home, credential_path)
     }
 
@@ -142,11 +126,7 @@ mod tests {
         let (move_home, credential_path) =
             setup_move_home("/save_credential_works_if_no_credential_file_exists");
         let _ = fs::remove_dir_all(&move_home);
-
-        let test_mode = Some(TestMode {
-            test_path: String::from("/save_credential_works_if_no_credential_file_exists"),
-        });
-        save_credential(String::from("test_token"), test_mode).unwrap();
+        MoveyLogin::save_credential(String::from("test_token"), move_home.clone()).unwrap();
 
         let contents = fs::read_to_string(&credential_path).expect("Unable to read file");
         let mut toml: Value = contents.parse().unwrap();
@@ -170,10 +150,7 @@ mod tests {
         let mut toml: Value = contents.parse().unwrap();
         assert!(toml.as_table_mut().unwrap().get_mut("registry").is_none());
 
-        let test_mode = Some(TestMode {
-            test_path: String::from("/save_credential_works_if_empty_credential_file_exists"),
-        });
-        save_credential(String::from("test_token"), test_mode).unwrap();
+        MoveyLogin::save_credential(String::from("test_token"), move_home.clone()).unwrap();
 
         let contents = fs::read_to_string(&credential_path).expect("Unable to read file");
         let mut toml: Value = contents.parse().unwrap();
@@ -204,10 +181,7 @@ mod tests {
         assert!(token.to_string().contains("old_test_token"));
         assert!(!token.to_string().contains("new_world"));
 
-        let test_mode = Some(TestMode {
-            test_path: String::from("/save_credential_works_if_token_field_exists"),
-        });
-        save_credential(String::from("new_world"), test_mode).unwrap();
+        MoveyLogin::save_credential(String::from("new_world"), move_home.clone()).unwrap();
 
         let contents = fs::read_to_string(&credential_path).expect("Unable to read file");
         let mut toml: Value = contents.parse().unwrap();
@@ -239,10 +213,7 @@ mod tests {
         let token = registry.as_table_mut().unwrap().get_mut("token").unwrap();
         assert!(!token.to_string().contains("test_token"));
 
-        let test_mode = Some(TestMode {
-            test_path: String::from("/save_credential_works_if_empty_token_field_exists"),
-        });
-        save_credential(String::from("test_token"), test_mode).unwrap();
+        MoveyLogin::save_credential(String::from("test_token"), move_home.clone()).unwrap();
 
         let contents = fs::read_to_string(&credential_path).expect("Unable to read file");
         let mut toml: Value = contents.parse().unwrap();
