+++ language/tools/move-cli/src/lib.rs
@@ -4,13 +4,12 @@
 
 use base::{
     build::Build, coverage::Coverage, disassemble::Disassemble, docgen::Docgen, errmap::Errmap,
-    info::Info, new::New, prove::Prove, test::Test,
+    info::Info, movey_login::MoveyLogin, new::New, prove::Prove, test::Test,
 };
 use move_package::BuildConfig;
 
 pub mod base;
 pub mod experimental;
-pub mod movey_login;
 pub mod sandbox;
 
 /// Default directory where saved Move resources live
@@ -93,10 +92,7 @@ pub enum Command {
         cmd: experimental::cli::ExperimentalCommand,
     },
     #[clap(name = "movey-login")]
-    MoveyLogin {
-        #[clap(long = "test-path")]
-        test_path: Option<String>,
-    },
+    MoveyLogin(MoveyLogin),
 }
 
 pub fn run_cli(
@@ -127,9 +123,7 @@ pub fn run_cli(
             &storage_dir,
         ),
         Command::Experimental { storage_dir, cmd } => cmd.handle_command(&move_args, &storage_dir),
-        Command::MoveyLogin { test_path } => {
-            movey_login::cli::handle_movey_login_commands(test_path)
-        }
+        Command::MoveyLogin(c) => c.execute(),
     }
 }
