--- language/tools/move-cli/src/movey_login/mod.rs
@@ -1,5 +1,0 @@
-// Copyright (c) The Diem Core Contributors
-// Copyright (c) The Move Contributors
-// SPDX-License-Identifier: Apache-2.0
-
-pub mod cli;
