+++ language/tools/move-cli/tests/cli_tests.rs
@@ -2,8 +2,8 @@
 // Copyright (c) The Move Contributors
 // SPDX-License-Identifier: Apache-2.0
 
-use move_cli::{movey_login::cli::MOVEY_API_KEY_PATH, sandbox::commands::test};
-use move_command_line_common::movey::MOVEY_URL;
+use move_cli::{base::movey_login::MOVEY_CREDENTIAL_PATH, sandbox::commands::test};
+use move_command_line_common::movey_constants::MOVEY_URL;
 #[cfg(unix)]
 use std::fs::File;
 use std::{env, fs, io::Write};
@@ -68,13 +68,9 @@ fn save_credential_works() {
     assert!(fs::read_to_string(&credential_path).is_err());
 
     match std::process::Command::new(cli_exe)
+        .env("MOVE_HOME", &move_home)
         .current_dir(".")
-        .args([
-            "movey-login",
-            "--test",
-            "--test-path",
-            "/save_credential_works",
-        ])
+        .args(["movey-login"])
         .stdin(Stdio::piped())
         .stdout(Stdio::piped())
         .spawn()
@@ -123,13 +119,9 @@ fn save_credential_fails_if_undeletable_credential_file_exists() {
     file.set_permissions(perms).unwrap();
 
     match std::process::Command::new(cli_exe)
+        .env("MOVE_HOME", &move_home)
         .current_dir(".")
-        .args([
-            "movey-login",
-            "--test",
-            "--test-path",
-            "/save_credential_fails_if_undeletable_credential_file_exists",
-        ])
+        .args(["movey-login"])
         .stdin(Stdio::piped())
         .stdout(Stdio::piped())
         .stderr(Stdio::piped())
@@ -171,11 +163,10 @@ fn save_credential_fails_if_undeletable_credential_file_exists() {
 fn setup_move_home(test_path: &str) -> (String, String) {
     let cwd = env::current_dir().unwrap();
     let mut move_home: String = String::from(cwd.to_string_lossy());
-    env::set_var("TEST_MOVE_HOME", &move_home);
     move_home.push_str(&test_path);
     let _ = fs::remove_dir_all(&move_home);
     fs::create_dir_all(&move_home).unwrap();
-    let credential_path = move_home.clone() + MOVEY_API_KEY_PATH;
+    let credential_path = move_home.clone() + MOVEY_CREDENTIAL_PATH;
     (move_home, credential_path)
 }
